Consolidate scenario and phases files
