+++ src/scenario/phases/common.ts
@@ -3,7 +3,7 @@
 // standalone via its own server endpoint, emitting the same ScenarioProgress
 // NDJSON envelope either way (the pattern established by buffer-passes).
 
-import type { ScenarioProgress } from '../scenario'
+import type { ScenarioConfig, ScenarioProgress } from '../scenario'
 
 // Phases report progress (and stream partial data) through this callback.
 export type PhaseEmit = (progress: ScenarioProgress) => void
@@ -18,28 +18,54 @@ export interface PhaseOpts {
 // Concurrent GraphQL requests per phase queue.
 export const PHASE_MAX_CONCURRENT_REQUESTS = 8
 
-// Phase identities for the progress plan and weights. 'buffers' covers all
-// four buffer passes as a single slice. 'stop-clusters' is the transfer-hub
-// pass (a separate stop query with nearby_stops neighbors).
-export type ScenarioPhaseName = 'feed-versions' | 'stops' | 'routes' | 'departures' | 'buffers' | 'stop-clusters' | 'flex-areas' | 'census-values'
+// The scenario phase registry: one entry per phase, in pipeline order. Adding
+// a phase is a single edit here — the name union, the order, the progress
+// weight, and the plan all derive from this array.
+//
+// `enabled` is the single source of truth for whether a config includes the
+// phase, driving both the emitted plan and fetchMain's execution gating, so the
+// two cannot drift. Routes, departures, and buffers execute inside the stops
+// block (they consume its ids), so their predicates must imply the stops
+// predicate. 'buffers' covers all four buffer passes as one slice;
+// 'stop-clusters' is the transfer-hub pass.
+//
+// Weights are relative progress-bar slices, normalized by the consumer over the
+// run's enabled phases. Rough cost ratios — tune from stage timings.
+const SCENARIO_PHASES = [
+  { name: 'feed-versions', weight: 1, enabled: () => true },
+  { name: 'stops', weight: 2, enabled: (c: ScenarioConfig) => c.includeFixedRoute !== false },
+  { name: 'routes', weight: 1, enabled: (c: ScenarioConfig) => c.includeFixedRoute !== false },
+  { name: 'departures', weight: 10, enabled: (c: ScenarioConfig) => c.includeFixedRoute !== false
+    && c.includeDepartures !== false },
+  { name: 'buffers', weight: 3, enabled: (c: ScenarioConfig) => c.includeFixedRoute !== false
+    && c.includeCensus !== false
+    && (c.stopBufferRadius ?? 0) > 0
+    && !!c.tableDatasetName },
+  { name: 'stop-clusters', weight: 2, enabled: (c: ScenarioConfig) => c.includeFixedRoute !== false
+    && (c.stopClusterDistance ?? 0) > 0 },
+  { name: 'flex-areas', weight: 1, enabled: (c: ScenarioConfig) => c.includeFlexAreas !== false },
+  { name: 'census-values', weight: 1, enabled: (c: ScenarioConfig) => c.includeCensus !== false
+    && !!c.tableDatasetName
+    && !!c.aggregateLayer },
+] as const satisfies readonly {
+  name: string
+  weight: number
+  enabled: (config: ScenarioConfig) => boolean
+}[]
+
+export type ScenarioPhaseName = typeof SCENARIO_PHASES[number]['name']
 
 // Pipeline ordering for phase plans and progress display.
-export const SCENARIO_PHASE_ORDER: ScenarioPhaseName[] = [
-  'feed-versions', 'stops', 'routes', 'departures', 'buffers', 'stop-clusters', 'flex-areas', 'census-values',
-]
+export const SCENARIO_PHASE_ORDER: ScenarioPhaseName[] = SCENARIO_PHASES.map(p => p.name)
 
-// Relative progress-bar weight per phase; the consumer normalizes over the
-// run's enabled plan. Rough cost ratios — tune from stage timings as data
-// accumulates. Equal weighting is the special case of all-1s.
-export const SCENARIO_PHASE_WEIGHTS: Record<ScenarioPhaseName, number> = {
-  'feed-versions': 1,
-  'stops': 2,
-  'routes': 1,
-  'departures': 10,
-  'buffers': 3,
-  'stop-clusters': 2,
-  'flex-areas': 1,
-  'census-values': 1,
+// `Object.fromEntries` widens its key type, so the cast restores exhaustiveness.
+export const SCENARIO_PHASE_WEIGHTS = Object.fromEntries(
+  SCENARIO_PHASES.map(p => [p.name, p.weight]),
+) as Record<ScenarioPhaseName, number>
+
+// The enabled phases for a scenario config, in pipeline order.
+export function scenarioPhasePlan (config: ScenarioConfig): ScenarioPhaseName[] {
+  return SCENARIO_PHASES.filter(p => p.enabled(config)).map(p => p.name)
 }
 
 // Final tick for a phase's progress slice. Also covers phases whose queue
