+++ src/scenario/phases/index.ts
@@ -2,13 +2,13 @@
 // config + GraphQL client + emit callback, streams ScenarioProgress events,
 // and returns the ids/context downstream phases need. ScenarioFetcher
 // composes them inline; server/api/scenario/* exposes each standalone.
-// (buffer-passes.ts is the seventh phase; it predates this directory.)
 
 // Re-export only common's public surface. The phase-contract plumbing
 // (PhaseEmit, PhaseOpts, phaseDone, PHASE_MAX_CONCURRENT_REQUESTS) stays
 // internal — phase modules import it directly from './common'.
 export {
   getSelectedDateRange,
+  scenarioPhasePlan,
   SCENARIO_PHASE_ORDER,
   SCENARIO_PHASE_WEIGHTS,
   type ScenarioPhaseName,
@@ -20,3 +20,5 @@ export * from './routes'
 export * from './departures'
 export * from './flex'
 export * from './census-values'
+export * from './buffer-passes'
+export * from './stop-clusters'
